還原 檔案 採用 Rev.1575 的原始檔
git-svn-id: http://www.castiello-tsai.org:8523/svn/L1JTW@1578 96aee97a-5c11-ed42-8518-d7f33e96f5f9

+++ src/net/l1j/server/WarTimeController.java
@@ -143,7 +143,7 @@ public class WarTimeController implements Runnable {
 								}
 							}
 							loc = L1CastleLocation.getGetBackLoc(castleId);
-							L1Teleport.teleport(pc, loc[0], loc[1], loc[2], 5, true);
+							L1Teleport.teleport(pc, loc[0], loc[1], (short) loc[2], 5, true);
 						}
 					}
 				}
