+++ src/net/l1j/server/command/executor/L1Loc.java
@@ -37,7 +37,7 @@ public class L1Loc implements L1CommandExecutor {
 		try {
 			int locx = pc.getX();
 			int locy = pc.getY();
-			int mapid = pc.getMapId();
+			short mapid = pc.getMapId();
 			int gab = L1WorldMap.getInstance().getMap(mapid).getOriginalTile(locx, locy);
 			String msg = String.format("座標 (%d, %d, %d) %d", locx, locy, mapid, gab);
 			pc.sendPackets(new S_SystemMessage(msg));
