+++ src/net/l1j/server/model/L1CastleLocation.java
@@ -54,7 +54,7 @@ public class L1CastleLocation {
 
 	private static final int KENT_TOWER_Y = 32774;
 
-	private static final int KENT_TOWER_MAP = 4;
+	private static final short KENT_TOWER_MAP = 4;
 
 	private static final int KENT_X1 = 33089;
 
@@ -64,16 +64,16 @@ public class L1CastleLocation {
 
 	private static final int KENT_Y2 = 32827;
 
-	private static final int KENT_MAP = 4;
+	private static final short KENT_MAP = 4;
 
-	private static final int KENT_INNER_CASTLE_MAP = 15;
+	private static final short KENT_INNER_CASTLE_MAP = 15;
 
 	// オークの森
 	private static final int OT_TOWER_X = 32800;
 
 	private static final int OT_TOWER_Y = 32290;
 
-	private static final int OT_TOWER_MAP = 4;
+	private static final short OT_TOWER_MAP = 4;
 
 	private static final int OT_X1 = 32750;
 
@@ -83,14 +83,14 @@ public class L1CastleLocation {
 
 	private static final int OT_Y2 = 32350;
 
-	private static final int OT_MAP = 4;
+	private static final short OT_MAP = 4;
 
 	// ウィンダウッド城
 	private static final int WW_TOWER_X = 32675;
 
 	private static final int WW_TOWER_Y = 33408;
 
-	private static final int WW_TOWER_MAP = 4;
+	private static final short WW_TOWER_MAP = 4;
 
 	private static final int WW_X1 = 32571;
 
@@ -100,16 +100,16 @@ public class L1CastleLocation {
 
 	private static final int WW_Y2 = 33460;
 
-	private static final int WW_MAP = 4;
+	private static final short WW_MAP = 4;
 
-	private static final int WW_INNER_CASTLE_MAP = 29;
+	private static final short WW_INNER_CASTLE_MAP = 29;
 
 	// ギラン城
 	private static final int GIRAN_TOWER_X = 33631;
 
 	private static final int GIRAN_TOWER_Y = 32678;
 
-	private static final int GIRAN_TOWER_MAP = 4;
+	private static final short GIRAN_TOWER_MAP = 4;
 
 	private static final int GIRAN_X1 = 33559;
 
@@ -119,16 +119,16 @@ public class L1CastleLocation {
 
 	private static final int GIRAN_Y2 = 32755;
 
-	private static final int GIRAN_MAP = 4;
+	private static final short GIRAN_MAP = 4;
 
-	private static final int GIRAN_INNER_CASTLE_MAP = 52;
+	private static final short GIRAN_INNER_CASTLE_MAP = 52;
 
 	// ハイネ城
 	private static final int HEINE_TOWER_X = 33524;
 
 	private static final int HEINE_TOWER_Y = 33396;
 
-	private static final int HEINE_TOWER_MAP = 4;
+	private static final short HEINE_TOWER_MAP = 4;
 
 	private static final int HEINE_X1 = 33458;
 
@@ -138,16 +138,16 @@ public class L1CastleLocation {
 
 	private static final int HEINE_Y2 = 33490;
 
-	private static final int HEINE_MAP = 4;
+	private static final short HEINE_MAP = 4;
 
-	private static final int HEINE_INNER_CASTLE_MAP = 64;
+	private static final short HEINE_INNER_CASTLE_MAP = 64;
 
 	// ドワーフ城
 	private static final int DOWA_TOWER_X = 32828;
 
 	private static final int DOWA_TOWER_Y = 32818;
 
-	private static final int DOWA_TOWER_MAP = 66;
+	private static final short DOWA_TOWER_MAP = 66;
 
 	private static final int DOWA_X1 = 32755;
 
@@ -157,14 +157,14 @@ public class L1CastleLocation {
 
 	private static final int DOWA_Y2 = 32920;
 
-	private static final int DOWA_MAP = 66;
+	private static final short DOWA_MAP = 66;
 
 	// アデン城
 	private static final int ADEN_TOWER_X = 34090;
 
 	private static final int ADEN_TOWER_Y = 33260;
 
-	private static final int ADEN_TOWER_MAP = 4;
+	private static final short ADEN_TOWER_MAP = 4;
 
 	private static final int ADEN_X1 = 34007;
 
@@ -174,9 +174,9 @@ public class L1CastleLocation {
 
 	private static final int ADEN_Y2 = 33332;
 
-	private static final int ADEN_MAP = 4;
+	private static final short ADEN_MAP = 4;
 
-	private static final int ADEN_INNER_CASTLE_MAP = 300;
+	private static final short ADEN_INNER_CASTLE_MAP = 300;
 
 	private static final int ADEN_SUB_TOWER1_X = 34057; // 青
 
@@ -199,7 +199,7 @@ public class L1CastleLocation {
 
 	private static final int DIAD_TOWER_Y = 32895;
 
-	private static final int DIAD_TOWER_MAP = 320;
+	private static final short DIAD_TOWER_MAP = 320;
 
 	private static final int DIAD_X1 = 32888;
 
@@ -209,9 +209,9 @@ public class L1CastleLocation {
 
 	private static final int DIAD_Y2 = 32953;
 
-	private static final int DIAD_MAP = 320;
+	private static final short DIAD_MAP = 320;
 
-	private static final int DIAD_INNER_CASTLE_MAP = 330;
+	private static final short DIAD_INNER_CASTLE_MAP = 330;
 
 	private static final Map<Integer, L1Location> _towers = new FastMap<Integer, L1Location>();
 
@@ -274,7 +274,7 @@ public class L1CastleLocation {
 	/**
 	 * ガーディアンタワー、クラウンの座標からcastle_idを返す
 	 */
-	public static int getCastleId(int locx, int locy, int mapid) {
+	public static int getCastleId(int locx, int locy, short mapid) {
 		return getCastleId(new L1Location(locx, locy, mapid));
 	}
 
@@ -317,7 +317,7 @@ public class L1CastleLocation {
 	/**
 	 * いずれかの戰爭エリア（旗內）かどうかチェック
 	 */
-	public static boolean checkInAllWarArea(int locx, int locy, int mapid) {
+	public static boolean checkInAllWarArea(int locx, int locy, short mapid) {
 		return checkInAllWarArea(new L1Location(locx, locy, mapid));
 	}
