+++ src/net/l1j/server/model/L1PetMatch.java
@@ -44,7 +44,7 @@ public class L1PetMatch {
 
 	public static final int MAX_PET_MATCH = 1;
 
-	private static final int[] PET_MATCH_MAPID = { 5125, 5131, 5132, 5133, 5134 };
+	private static final short[] PET_MATCH_MAPID = { 5125, 5131, 5132, 5133, 5134 };
 
 	private String[] _pc1Name = new String[MAX_PET_MATCH];
 	private String[] _pc2Name = new String[MAX_PET_MATCH];
@@ -241,7 +241,7 @@ public class L1PetMatch {
 					pet.deleteMe();
 				}
 			}
-			L1Teleport.teleport(pc1, 32630, 32744, 4, 4, true);
+			L1Teleport.teleport(pc1, 32630, 32744, (short) 4, 4, true);
 		}
 		_pc1Name[petMatchNo] = null;
 		_pet1[petMatchNo] = null;
@@ -256,7 +256,7 @@ public class L1PetMatch {
 					pet.deleteMe();
 				}
 			}
-			L1Teleport.teleport(pc2, 32630, 32744, 4, 4, true);
+			L1Teleport.teleport(pc2, 32630, 32744, (short) 4, 4, true);
 		}
 		_pc2Name[petMatchNo] = null;
 		_pet2[petMatchNo] = null;
