+++ src/net/l1j/server/model/instance/L1CrownInstance.java
@@ -122,7 +122,7 @@ public class L1CrownInstance extends L1NpcInstance {
 					loc = L1CastleLocation.getGetBackLoc(castle_id);
 					int locx = loc[0];
 					int locy = loc[1];
-					int mapid = loc[2];
+					short mapid = (short) loc[2];
 					L1Teleport.teleport(pc, locx, locy, mapid, 5, true);
 				}
 			}
