+++ src/net/l1j/server/model/instance/L1TeleporterInstance.java
@@ -463,12 +463,12 @@ public class L1TeleporterInstance extends L1NpcInstance {
 					// 3マス以內のPc
 					for (L1PcInstance otherPc : L1World.getInstance().getVisiblePlayer(player, 3)) {
 						if (otherPc.getClanid() == player.getClanid() && otherPc.getId() != player.getId()) {
-							L1Teleport.teleport(otherPc, 32740, 32800, 217, 5, true);
+							L1Teleport.teleport(otherPc, 32740, 32800, (short) 217, 5, true);
 						}
 					}
-					L1Teleport.teleport(player, 32740, 32800, 217, 5, true);
+					L1Teleport.teleport(player, 32740, 32800, (short) 217, 5, true);
 				} else if (action.equalsIgnoreCase("teleport mage-quest-dungen")) { // 試練のダンジョン（ウィザードLv30クエスト）
-					L1Teleport.teleport(player, 32791, 32788, 201, 5, true);
+					L1Teleport.teleport(player, 32791, 32788, (short) 201, 5, true);
 				} else if (action.equalsIgnoreCase("teleport 29")) { // ラムダ
 					L1PcInstance kni = null;
 					L1PcInstance elf = null;
@@ -494,15 +494,15 @@ public class L1TeleporterInstance extends L1NpcInstance {
 						}
 					}
 					if (kni != null && elf != null && wiz != null) { // 全クラス揃っている
-						L1Teleport.teleport(player, 32723, 32850, 2000, 2, true);
-						L1Teleport.teleport(kni, 32750, 32851, 2000, 6, true);
-						L1Teleport.teleport(elf, 32878, 32980, 2000, 6, true);
-						L1Teleport.teleport(wiz, 32876, 33003, 2000, 0, true);
+						L1Teleport.teleport(player, 32723, 32850, (short) 2000, 2, true);
+						L1Teleport.teleport(kni, 32750, 32851, (short) 2000, 6, true);
+						L1Teleport.teleport(elf, 32878, 32980, (short) 2000, 6, true);
+						L1Teleport.teleport(wiz, 32876, 33003, (short) 2000, 0, true);
 						TeleportDelyTimer timer = new TeleportDelyTimer();
 						ThreadPoolManager.getInstance().execute(timer);
 					}
 				} else if (action.equalsIgnoreCase("teleport barlog")) { // 古代人（Lv50クエスト古代の空間2F）
-					L1Teleport.teleport(player, 32755, 32844, 2002, 5, true);
+					L1Teleport.teleport(player, 32755, 32844, (short) 2002, 5, true);
 					TeleportDelyTimer timer = new TeleportDelyTimer();
 					ThreadPoolManager.getInstance().execute(timer);
 				}
